refactor: improve code quality and request typing and remove unsafe casts

+++ src/schemas/product.schema.ts
@@ -2,7 +2,7 @@ import z from "zod";
 
 export const CreateProductSchema = z.object({
   productName: z.string(),
-  productPrice: z.coerce.number().int().positive(),
+  productPrice: z.coerce.number().positive(),
   stock: z.coerce.number().int().positive(),
 });
